irker-svnpoller: Deliver to multiple channels at once
irkerd can handle messages set to be delivered to a channel list. This
allows us to contact irkerd just once with the list, instead of multiple
times iterating over each channel URL.

At this time this doesn't seem to make a big difference within the
irkerd<->IRC portion of the pipeline, but this approach should allow
irker-svnpoller to take advantage of any future irkerd optimizations.